Change video_length column to decimal

+++ database/migrations/2021_03_03_173015_add_length_to_video.php
@@ -14,7 +14,7 @@ class AddLengthToVideo extends Migration
     public function up()
     {
         Schema::table('videos', function (Blueprint $table) {
-            $table->integer('video_length');
+            $table->decimal('video_length');
         });
     }
